IRIDA-10: Has been changed InlineKeyboardService.
