+++ src/main/java/com/github/iridatelegrambot/service/buttons/MenuButtonsService.java
@@ -1,6 +1,5 @@
 package com.github.iridatelegrambot.service.buttons;
 
-import org.telegram.telegrambots.meta.api.objects.replykeyboard.InlineKeyboardMarkup;
 import org.telegram.telegrambots.meta.api.objects.replykeyboard.ReplyKeyboardMarkup;
 
 public interface MenuButtonsService {
