--- src/main/java/com/github/iridatelegrambot/service/senders/SendAdminMenuService.java
@@ -1,10 +1,0 @@
-package com.github.iridatelegrambot.service.senders;
-
-public interface SendAdminMenuService{
-
-    void sendAdminMenu(Long chatId, String message,Integer idMessage);
-
-    void sendAdminSetStatus(Long chatId, boolean status, String message, Integer messageId);
-
-    void sendUsersForAdmin(Long chatId, String message, Integer messageId);
-}
