+++ src/main/java/com/github/iridatelegrambot/service/senders/SendDocumentServiceImpl.java
@@ -4,27 +4,52 @@ import com.github.iridatelegrambot.entity.Invoice;
 import com.github.iridatelegrambot.entity.Order;
 import com.github.iridatelegrambot.service.InvoiceService;
 import com.github.iridatelegrambot.service.OrderService;
-import com.github.iridatelegrambot.service.buttons.InlineKeyboardService;
+import com.github.iridatelegrambot.service.buttons.InlineDocumentButtonService;
 import com.github.iridatelegrambot.service.statuswait.WaitDocument;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
+import org.telegram.telegrambots.meta.api.objects.replykeyboard.InlineKeyboardMarkup;
 
 import java.util.Optional;
-
 @Service
-public class SendWithOrderServiceImpl implements SendWithOrderService {
+public class SendDocumentServiceImpl implements SendDocumentService {
     private final SendMessageService sendMessageService;
-    private final InlineKeyboardService inlineKeyboardService;
+    private final InlineDocumentButtonService inlineKeyboardService;
     private final OrderService orderService;
     private final InvoiceService invoiceService;
 
     @Autowired
-    public SendWithOrderServiceImpl(SendMessageService sendMessageService, InlineKeyboardService inlineKeyboardService,
-                                    OrderService orderService, InvoiceService invoiceService) {
+    public SendDocumentServiceImpl(SendMessageService sendMessageService, InlineDocumentButtonService inlineKeyboardService,
+        OrderService orderService, InvoiceService invoiceService){
         this.sendMessageService = sendMessageService;
         this.inlineKeyboardService = inlineKeyboardService;
         this.orderService = orderService;
         this.invoiceService = invoiceService;
+    }
+
+    @Override
+    public void sendListCityForOrder(Optional<Order> orderOptional, Long chatId){
+        if(orderOptional.isEmpty()){
+            sendMessageService.sendMessage(chatId.toString(),
+                    "Повторите попытку.\nСообщение не содержит номера или такой номер уже есть.");
+            return;
+        }
+
+        Order order = orderOptional.get();
+        InlineKeyboardMarkup inlineKeyboardMarkup = inlineKeyboardService.cityButtons(order);
+        sendMessageService.sendMessage(chatId.toString(),"Введите город который сделал заказ:",inlineKeyboardMarkup);
+    }
+    @Override
+    public void sendListCityForInvoice(Optional<Invoice> invoiceOptional, Long chatId){
+        if (invoiceOptional.isEmpty()){
+            sendMessageService.sendMessage(chatId.toString(),
+                    "Повторите попытку.\nСообщение не содержит номера или такой номер уже есть.");
+            return;
+        }
+
+        Invoice invoice = invoiceOptional.get();
+        InlineKeyboardMarkup inlineKeyboardMarkup = inlineKeyboardService.cityButtons(invoice);
+        sendMessageService.sendMessage(chatId.toString(),"Введите город из которого была накладная:",inlineKeyboardMarkup);
     }
 
     @Override
