+++ src/main/java/com/github/iridatelegrambot/service/senders/SendInviteForAdminServiceImpl.java
@@ -1,15 +1,15 @@
 package com.github.iridatelegrambot.service.senders;
 
-import com.github.iridatelegrambot.service.buttons.InlineKeyboardService;
+import com.github.iridatelegrambot.service.buttons.InlineAdminButtonService;
 import org.springframework.stereotype.Service;
 
 @Service
 public class SendInviteForAdminServiceImpl implements SendInviteForAdminService{
 
     private final SendMessageService sendMessageService;
-    private final InlineKeyboardService inlineKeyboardService;
+    private final InlineAdminButtonService inlineKeyboardService;
 
-    public SendInviteForAdminServiceImpl(SendMessageService sendMessageService, InlineKeyboardService service) {
+    public SendInviteForAdminServiceImpl(SendMessageService sendMessageService, InlineAdminButtonService service) {
         this.sendMessageService = sendMessageService;
         this.inlineKeyboardService = service;
     }
