+++ src/test/java/com/github/iridatelegrambot/test/command/AbstractCommandTest.java
@@ -3,28 +3,18 @@ package com.github.iridatelegrambot.test.command;
 import com.github.iridatelegrambot.command.Command;
 import com.github.iridatelegrambot.service.senders.CommandSenderService;
 import com.github.iridatelegrambot.service.senders.CommandSenderServiceImpl;
-import com.github.iridatelegrambot.service.senders.SendMessageServiceImpl;
 import com.github.iridatelegrambot.service.UserTelegramService;
-import com.github.iridatelegrambot.bot.IridaBot;
-import com.github.iridatelegrambot.service.buttons.InlineKeyboardService;
-import com.github.iridatelegrambot.service.buttons.MenuButtonsService;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
-import org.mockito.Mock;
 import org.mockito.Mockito;
-import org.telegram.telegrambots.meta.api.methods.send.SendMessage;
 import org.telegram.telegrambots.meta.api.objects.Chat;
 import org.telegram.telegrambots.meta.api.objects.Message;
 import org.telegram.telegrambots.meta.api.objects.Update;
-import org.telegram.telegrambots.meta.exceptions.TelegramApiException;
 
 abstract class AbstractCommandTest {
     protected UserTelegramService mUserTelegramService = Mockito.mock(UserTelegramService.class);
-    protected IridaBot mIridaBot = Mockito.mock(IridaBot.class);
-    protected InlineKeyboardService mInlineKeyboardService = Mockito.mock(InlineKeyboardService.class);
-    protected MenuButtonsService menuButtonsService = Mockito.mock(MenuButtonsService.class);
-    protected SendMessageServiceImpl sendMessageService = new SendMessageServiceImpl(mIridaBot);
     protected CommandSenderService commandSenderService = Mockito.mock(CommandSenderServiceImpl.class);
+    protected Long chatId = 12345678L;
 
 
     abstract String getCommandName();
@@ -33,13 +23,10 @@ abstract class AbstractCommandTest {
 
     abstract Command getCommand();
 
-    @Test
-    void shouldExecuteCommand() throws TelegramApiException {
-        //given
-        Long chatId = 12345678L;
+    abstract void shouldExecuteCommand();
 
+    Update createUpdate() {
         Update update = new Update();
-
         Message message = Mockito.mock(Message.class);
         Mockito.when(message.getChatId()).thenReturn(chatId);
         Mockito.when(message.getText()).thenReturn(getCommandName());
@@ -50,16 +37,7 @@ abstract class AbstractCommandTest {
         message.setChat(chat);
         update.setMessage(message);
 
-        SendMessage sendMessage = new SendMessage();
-        sendMessage.setChatId(chatId.toString());
-        sendMessage.setText(getCommandMessage());
-        sendMessage.enableHtml(true);
-
-        //when
-        getCommand().execute(update);
-
-        //then
-        Mockito.verify(mIridaBot).execute(sendMessage);
+        return update;
     }
 
     @Test
