+++ src/test/java/com/github/iridatelegrambot/test/command/NoCommandTest.java
@@ -1,6 +1,8 @@
 package com.github.iridatelegrambot.test.command;
 import com.github.iridatelegrambot.command.Command;
 import com.github.iridatelegrambot.command.NoCommand;
+import org.junit.jupiter.api.Test;
+import org.mockito.Mockito;
 
 import static com.github.iridatelegrambot.command.CommandName.NO;
 import static com.github.iridatelegrambot.command.NoCommand.NO_MESSAGE;
@@ -19,5 +21,13 @@ public class NoCommandTest extends AbstractCommandTest {
     @Override
     Command getCommand() {
         return new NoCommand(commandSenderService);
+    }
+    @Test
+    @Override
+    void shouldExecuteCommand() {
+        //when
+        getCommand().execute(createUpdate());
+        //then
+        Mockito.verify(commandSenderService).sendMessage(chatId.toString(),getCommandMessage());
     }
 }
