+++ src/test/java/com/github/iridatelegrambot/test/command/ShowActiveOrdersCommandTest.java
@@ -2,6 +2,8 @@ package com.github.iridatelegrambot.test.command;
 
 import com.github.iridatelegrambot.command.Command;
 import com.github.iridatelegrambot.command.ShowActiveOrdersCommand;
+import org.junit.jupiter.api.Test;
+import org.mockito.Mockito;
 
 import static com.github.iridatelegrambot.command.CommandName.SHOWACTIVEORDER;
 import static com.github.iridatelegrambot.command.ShowActiveOrdersCommand.SHOW_ACTIVE_ORDERS_MESSAGE;
@@ -20,5 +22,13 @@ public class ShowActiveOrdersCommandTest extends AbstractCommandTest{
     @Override
     Command getCommand() {
         return new ShowActiveOrdersCommand(commandSenderService);
+    }
+    @Test
+    @Override
+    void shouldExecuteCommand() {
+        //when
+        getCommand().execute(createUpdate());
+        //then
+        Mockito.verify(commandSenderService).sendActiveOrders(chatId,getCommandMessage());
     }
 }
