+++ src/test/java/com/github/iridatelegrambot/test/command/StopCommandTest.java
@@ -2,6 +2,8 @@ package com.github.iridatelegrambot.test.command;
 
 import com.github.iridatelegrambot.command.Command;
 import com.github.iridatelegrambot.command.StopCommand;
+import org.junit.jupiter.api.Test;
+import org.mockito.Mockito;
 
 import static com.github.iridatelegrambot.command.CommandName.STOP;
 import static com.github.iridatelegrambot.command.StopCommand.STOP_MESSAGE;
@@ -20,5 +22,13 @@ public class StopCommandTest extends AbstractCommandTest {
     @Override
     Command getCommand() {
         return new StopCommand(commandSenderService,mUserTelegramService);
+    }
+    @Test
+    @Override
+    void shouldExecuteCommand() {
+        //when
+        getCommand().execute(createUpdate());
+        //then
+        Mockito.verify(commandSenderService).sendMessage(chatId.toString(),getCommandMessage());
     }
 }
