+++ src/test/java/com/github/iridatelegrambot/test/command/UnknownCommandTest.java
@@ -6,6 +6,7 @@ import com.github.iridatelegrambot.command.CommandName;
 import com.github.iridatelegrambot.command.UnknownCommand;
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
+import org.mockito.Mockito;
 
 import static com.github.iridatelegrambot.command.UnknownCommand.UNKNOWN_MESSAGE;
 
@@ -29,5 +30,14 @@ public class UnknownCommandTest extends AbstractCommandTest{
     @Override
     void shouldGetProperlyCommand(){
         Assertions.assertEquals(getCommandName(),getCommand().getCommand().getCommandName());
+    }
+
+    @Test
+    @Override
+    void shouldExecuteCommand() {
+        //when
+        getCommand().execute(createUpdate());
+        //then
+        Mockito.verify(commandSenderService).sendMessage(chatId.toString(),getCommandMessage());
     }
 }
